feat(draw/sw): make lv_draw_mask_rect_dsc_t public
