+++ lvgl_private.h
@@ -35,7 +35,6 @@ extern "C" {
 #include "src/draw/lv_draw_label_private.h"
 #include "src/draw/lv_draw_vector_private.h"
 #include "src/draw/lv_draw_buf_private.h"
-#include "src/draw/lv_draw_mask_private.h"
 #include "src/draw/sw/lv_draw_sw_private.h"
 #include "src/draw/sw/lv_draw_sw_mask_private.h"
 #include "src/draw/sw/blend/lv_draw_sw_blend_private.h"
