+++ src/draw/lv_draw_mask.c
@@ -6,7 +6,7 @@
 /*********************
  *      INCLUDES
  *********************/
-#include "lv_draw_mask_private.h"
+#include "lv_draw_mask.h"
 #include "lv_draw_private.h"
 #include "../core/lv_refr.h"
 #include "../misc/lv_math.h"
