+++ src/draw/lv_draw_mask.h
@@ -13,6 +13,7 @@ extern "C" {
 /*********************
  *      INCLUDES
  *********************/
+#include "lv_draw.h"
 #include "../misc/lv_color.h"
 #include "../misc/lv_area.h"
 #include "../misc/lv_style.h"
@@ -20,6 +21,24 @@ extern "C" {
 /*********************
  *      DEFINES
  *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+struct _lv_draw_mask_rect_dsc_t {
+    lv_draw_dsc_base_t base;
+
+    /**The area to mask.*/
+    lv_area_t area;
+
+    /**The radius of masking*/
+    int32_t radius;
+
+    /**0: clear the content out of the `area`.
+     * 1: don't touch the area out of `area`*/
+    uint32_t keep_outside : 1;
+};
 
 /**********************
  * GLOBAL PROTOTYPES
