+++ src/draw/nanovg/lv_draw_nanovg_mask_rect.c
@@ -11,7 +11,7 @@
 
 #if LV_USE_DRAW_NANOVG
 
-#include "../../draw/lv_draw_mask_private.h"
+#include "../../draw/lv_draw_mask.h"
 #include "lv_nanovg_utils.h"
 
 /*********************
