+++ src/draw/nema_gfx/lv_draw_nema_gfx.h
@@ -50,7 +50,7 @@ extern "C" {
 #include "../lv_draw_image_private.h"
 #include "../lv_image_decoder_private.h"
 #include "../lv_draw_label_private.h"
-#include "../lv_draw_mask_private.h"
+#include "../lv_draw_mask.h"
 #include "../lv_draw_rect_private.h"
 #include "../lv_draw_triangle_private.h"
 #include "../lv_draw_vector_private.h"
