+++ src/widgets/bar/lv_bar.c
@@ -8,7 +8,7 @@
  *********************/
 #include "lv_bar_private.h"
 #include "../../misc/lv_area_private.h"
-#include "../../draw/lv_draw_mask_private.h"
+#include "../../draw/lv_draw_mask.h"
 #include "../../core/lv_obj_private.h"
 #include "../../core/lv_obj_class_private.h"
 #if LV_USE_BAR != 0
